backup_working_version_2025-07-09_13-06-12/requirements.txt
Pillow>=10.1.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
numpy>=1.21.0
tkinterdnd2>=0.3.0 